Implemented Power State feature
After the backup operation has been successfully completed, the program
is now able to send the following signals to the Operating System:
Suspend, Hibernation, and  Shutdown power states.  The grace time period
for some of these commands is 5 minutes, which the user is given
instructions (right on the screen before t he command is issued) to
abort the count time.  Of course, the user can disable this feature
freely.