Added Legacy build
Despite being unsupported, I am going to include this build in the
official repo if incase others ever need it for specific reasons.  Do
note for those that use this build for whatever reason are subjected to
its environment,  limitations, and bugs.  I must stress that this build
is not maintained and is considered unsupported.  Use at your own risk.