Revised the %core.Date% to use the English region time and date formatting
When I was developing this program, I used the German Time and Date region formatting.  With using the German formatting, the %DATE% or 'DATE /T' would provide this formatting: DD.MM.YY.  Yet, the English formatting is: DDD MM/DD/YYYY.  Because of this inconsistency, and no real easy way to quickly fix this, the regional formatting will have to be set as 'English'.  With that, the %core.Date% will contain: Day, Month, and Year - but still following the user's English regional settings.